+++ smsdk_config.h
@@ -39,9 +39,9 @@
 
 #define SMEXT_CONF_NAME            "DODHooks"
 #define SMEXT_CONF_DESCRIPTION     "SourceMod extension with Detours & Natives for Day of Defeat: Source"
-#define SMEXT_CONF_VERSION         "1.6.2"
+#define SMEXT_CONF_VERSION         "1.6.3"
 #define SMEXT_CONF_AUTHOR          "Andersso, ChesterSmitty, Apfelwurm, DNA-styx, Kittenks"
-#define SMEXT_CONF_URL             "https://github.com/kittenks/dodhooks"
+#define SMEXT_CONF_URL             "https://github.com/sppluginteam/dodhooks-beta"
 #define SMEXT_CONF_LOGTAG          "DODHOOKS"
 #define SMEXT_CONF_LICENSE         "GPLv2"
 #define SMEXT_CONF_DATESTRING      __DATE__